Add a bunch of hardenings/fixed on top of 6.x redesign
The 6.x multi-process design is clear improvement, especially since
crash of the "gfx" process no longer unlocks the screen. But it has also
few downsides. The most noticeable one is the "gfx" process (the one
responsible for blanking the screen) is not running all the
time, but only when screen is locked. Especially, when DPMS kicks in
earlier, the "gfx" process starts when the output is in power saving
mode already, and standard screen blanking doesn't work anymore. Patches
1 and 3 deal with this issue.
Besides that, handle few corner cases better - see patch descriptions
for details.

The below "Fixes" should really belong to update to 6.x, but without
this set of patches, it isn't suitable for Qubes, so lets stick issue
reference here instead.
Fixes QubesOS/qubes-issues#4394